+++ apps/console/src/hooks/graph/__generated__/ContinualImprovementRegistryGraphNodeQuery.graphql.ts
@@ -1,5 +1,5 @@
 /**
- * @generated SignedSource<<6ac0e2eaa895ed1c2409e524e0333ab6>>
+ * @generated SignedSource<<9992eeb0b74e7015c5f4d3520a87a402>>
  * @lightSyntaxTransform
  * @nogrep
  */
@@ -16,14 +16,6 @@ export type ContinualImprovementRegistryGraphNodeQuery$variables = {
 };
 export type ContinualImprovementRegistryGraphNodeQuery$data = {
   readonly node: {
-    readonly audit?: {
-      readonly framework: {
-        readonly id: string;
-        readonly name: string;
-      };
-      readonly id: string;
-      readonly name: string | null | undefined;
-    };
     readonly createdAt?: any;
     readonly description?: string | null | undefined;
     readonly id?: string;
@@ -37,7 +29,9 @@ export type ContinualImprovementRegistryGraphNodeQuery$data = {
     };
     readonly priority?: ContinualImprovementRegistriesPriority;
     readonly referenceId?: string;
+    readonly snapshotId?: string | null | undefined;
     readonly source?: string | null | undefined;
+    readonly sourceId?: string | null | undefined;
     readonly status?: ContinualImprovementRegistriesStatus;
     readonly targetDate?: any | null | undefined;
     readonly updatedAt?: any;
@@ -74,79 +68,59 @@ v3 = {
   "alias": null,
   "args": null,
   "kind": "ScalarField",
-  "name": "referenceId",
+  "name": "snapshotId",
   "storageKey": null
 },
 v4 = {
   "alias": null,
   "args": null,
   "kind": "ScalarField",
-  "name": "description",
+  "name": "sourceId",
   "storageKey": null
 },
 v5 = {
   "alias": null,
   "args": null,
   "kind": "ScalarField",
-  "name": "source",
+  "name": "referenceId",
   "storageKey": null
 },
 v6 = {
   "alias": null,
   "args": null,
   "kind": "ScalarField",
-  "name": "targetDate",
+  "name": "description",
   "storageKey": null
 },
 v7 = {
   "alias": null,
   "args": null,
   "kind": "ScalarField",
-  "name": "status",
+  "name": "source",
   "storageKey": null
 },
 v8 = {
   "alias": null,
   "args": null,
   "kind": "ScalarField",
-  "name": "priority",
+  "name": "targetDate",
   "storageKey": null
 },
 v9 = {
   "alias": null,
   "args": null,
   "kind": "ScalarField",
-  "name": "name",
+  "name": "status",
   "storageKey": null
 },
-v10 = [
-  (v2/*: any*/),
-  (v9/*: any*/)
-],
-v11 = {
+v10 = {
   "alias": null,
   "args": null,
-  "concreteType": "Audit",
-  "kind": "LinkedField",
-  "name": "audit",
-  "plural": false,
-  "selections": [
-    (v2/*: any*/),
-    (v9/*: any*/),
-    {
-      "alias": null,
-      "args": null,
-      "concreteType": "Framework",
-      "kind": "LinkedField",
-      "name": "framework",
-      "plural": false,
-      "selections": (v10/*: any*/),
-      "storageKey": null
-    }
-  ],
+  "kind": "ScalarField",
+  "name": "priority",
   "storageKey": null
 },
-v12 = {
+v11 = {
   "alias": null,
   "args": null,
   "concreteType": "People",
@@ -165,24 +139,33 @@ v12 = {
   ],
   "storageKey": null
 },
-v13 = {
+v12 = {
   "alias": null,
   "args": null,
   "concreteType": "Organization",
   "kind": "LinkedField",
   "name": "organization",
   "plural": false,
-  "selections": (v10/*: any*/),
+  "selections": [
+    (v2/*: any*/),
+    {
+      "alias": null,
+      "args": null,
+      "kind": "ScalarField",
+      "name": "name",
+      "storageKey": null
+    }
+  ],
   "storageKey": null
 },
-v14 = {
+v13 = {
   "alias": null,
   "args": null,
   "kind": "ScalarField",
   "name": "createdAt",
   "storageKey": null
 },
-v15 = {
+v14 = {
   "alias": null,
   "args": null,
   "kind": "ScalarField",
@@ -214,11 +197,12 @@ return {
               (v6/*: any*/),
               (v7/*: any*/),
               (v8/*: any*/),
+              (v9/*: any*/),
+              (v10/*: any*/),
               (v11/*: any*/),
               (v12/*: any*/),
               (v13/*: any*/),
-              (v14/*: any*/),
-              (v15/*: any*/)
+              (v14/*: any*/)
             ],
             "type": "ContinualImprovementRegistry",
             "abstractKey": null
@@ -261,11 +245,12 @@ return {
               (v6/*: any*/),
               (v7/*: any*/),
               (v8/*: any*/),
+              (v9/*: any*/),
+              (v10/*: any*/),
               (v11/*: any*/),
               (v12/*: any*/),
               (v13/*: any*/),
-              (v14/*: any*/),
-              (v15/*: any*/)
+              (v14/*: any*/)
             ],
             "type": "ContinualImprovementRegistry",
             "abstractKey": null
@@ -276,16 +261,16 @@ return {
     ]
   },
   "params": {
-    "cacheID": "c3195d70369df1a1ca2a73fdfef37039",
+    "cacheID": "dc67877ad1fe94bfeb50490a9ca6692b",
     "id": null,
     "metadata": {},
     "name": "ContinualImprovementRegistryGraphNodeQuery",
     "operationKind": "query",
-    "text": "query ContinualImprovementRegistryGraphNodeQuery(\n  $continualImprovementRegistryId: ID!\n) {\n  node(id: $continualImprovementRegistryId) {\n    __typename\n    ... on ContinualImprovementRegistry {\n      id\n      referenceId\n      description\n      source\n      targetDate\n      status\n      priority\n      audit {\n        id\n        name\n        framework {\n          id\n          name\n        }\n      }\n      owner {\n        id\n        fullName\n      }\n      organization {\n        id\n        name\n      }\n      createdAt\n      updatedAt\n    }\n    id\n  }\n}\n"
+    "text": "query ContinualImprovementRegistryGraphNodeQuery(\n  $continualImprovementRegistryId: ID!\n) {\n  node(id: $continualImprovementRegistryId) {\n    __typename\n    ... on ContinualImprovementRegistry {\n      id\n      snapshotId\n      sourceId\n      referenceId\n      description\n      source\n      targetDate\n      status\n      priority\n      owner {\n        id\n        fullName\n      }\n      organization {\n        id\n        name\n      }\n      createdAt\n      updatedAt\n    }\n    id\n  }\n}\n"
   }
 };
 })();
 
-(node as any).hash = "fe3c3ff44245eb13ea465ae5809ef5bd";
+(node as any).hash = "659dd712ecf01765f6d5ef0f9d7d2a08";
 
 export default node;