+++ apps/console/src/routes/continualImprovementRegistryRoutes.ts
@@ -9,14 +9,40 @@ export const continualImprovementRegistryRoutes = [
   {
     path: "continual-improvement-registries",
     fallback: PageSkeleton,
-    queryLoader: ({ organizationId }: { organizationId: string }) =>
-      loadQuery(relayEnvironment, continualImprovementRegistriesQuery, { organizationId }),
+    queryLoader: (params: Record<string, string>) =>
+      loadQuery(relayEnvironment, continualImprovementRegistriesQuery, {
+        organizationId: params.organizationId,
+        snapshotId: null
+      }),
+    Component: lazy(
+      () => import("/pages/organizations/continualImprovementRegistries/ContinualImprovementRegistriesPage")
+    ),
+  },
+  {
+    path: "snapshots/:snapshotId/continual-improvement-registries",
+    fallback: PageSkeleton,
+    queryLoader: (params: Record<string, string>) =>
+      loadQuery(relayEnvironment, continualImprovementRegistriesQuery, {
+        organizationId: params.organizationId,
+        snapshotId: params.snapshotId
+      }),
     Component: lazy(
       () => import("/pages/organizations/continualImprovementRegistries/ContinualImprovementRegistriesPage")
     ),
   },
   {
     path: "continual-improvement-registries/:registryId",
+    fallback: PageSkeleton,
+    queryLoader: (params: Record<string, string>) =>
+      loadQuery(relayEnvironment, continualImprovementRegistryNodeQuery, {
+        continualImprovementRegistryId: params.registryId
+      }),
+    Component: lazy(
+      () => import("/pages/organizations/continualImprovementRegistries/ContinualImprovementRegistryDetailsPage")
+    ),
+  },
+  {
+    path: "snapshots/:snapshotId/continual-improvement-registries/:registryId",
     fallback: PageSkeleton,
     queryLoader: (params: Record<string, string>) =>
       loadQuery(relayEnvironment, continualImprovementRegistryNodeQuery, {