+++ pkg/coredata/continual_improvement_registries.go
@@ -32,12 +32,13 @@ type (
 		OrganizationID gid.GID                                `db:"organization_id"`
 		ReferenceID    string                                 `db:"reference_id"`
 		Description    *string                                `db:"description"`
-		AuditID        gid.GID                                `db:"audit_id"`
 		Source         *string                                `db:"source"`
 		OwnerID        gid.GID                                `db:"owner_id"`
 		TargetDate     *time.Time                             `db:"target_date"`
 		Status         ContinualImprovementRegistriesStatus   `db:"status"`
 		Priority       ContinualImprovementRegistriesPriority `db:"priority"`
+		SnapshotID     *gid.GID                               `db:"snapshot_id"`
+		SourceID       *gid.GID                               `db:"source_id"`
 		CreatedAt      time.Time                              `db:"created_at"`
 		UpdatedAt      time.Time                              `db:"updated_at"`
 	}
@@ -74,12 +75,13 @@ SELECT
 	organization_id,
 	reference_id,
 	description,
-	audit_id,
 	source,
 	owner_id,
 	target_date,
 	status,
 	priority,
+	snapshot_id,
+	source_id,
 	created_at,
 	updated_at
 FROM
@@ -115,6 +117,7 @@ func (cirs *ContinualImprovementRegistries) CountByOrganizationID(
 	conn pg.Conn,
 	scope Scoper,
 	organizationID gid.GID,
+	filter *ContinualImprovementRegistryFilter,
 ) (int, error) {
 	q := `
 SELECT
@@ -124,12 +127,14 @@ FROM
 WHERE
 	%s
 	AND organization_id = @organization_id
+	AND %s
 `
 
-	q = fmt.Sprintf(q, scope.SQLFragment())
+	q = fmt.Sprintf(q, scope.SQLFragment(), filter.SQLFragment())
 
 	args := pgx.StrictNamedArgs{"organization_id": organizationID}
 	maps.Copy(args, scope.SQLArguments())
+	maps.Copy(args, filter.SQLArguments())
 
 	row := conn.QueryRow(ctx, q, args)
 
@@ -148,6 +153,7 @@ func (cirs *ContinualImprovementRegistries) LoadByOrganizationID(
 	scope Scoper,
 	organizationID gid.GID,
 	cursor *page.Cursor[ContinualImprovementRegistriesOrderField],
+	filter *ContinualImprovementRegistryFilter,
 ) error {
 	q := `
 SELECT
@@ -155,12 +161,13 @@ SELECT
 	organization_id,
 	reference_id,
 	description,
-	audit_id,
 	source,
 	owner_id,
 	target_date,
 	status,
 	priority,
+	snapshot_id,
+	source_id,
 	created_at,
 	updated_at
 FROM
@@ -169,12 +176,14 @@ WHERE
 	%s
 	AND organization_id = @organization_id
 	AND %s
+	AND %s
 `
 
-	q = fmt.Sprintf(q, scope.SQLFragment(), cursor.SQLFragment())
+	q = fmt.Sprintf(q, scope.SQLFragment(), filter.SQLFragment(), cursor.SQLFragment())
 
 	args := pgx.StrictNamedArgs{"organization_id": organizationID}
 	maps.Copy(args, scope.SQLArguments())
+	maps.Copy(args, filter.SQLArguments())
 	maps.Copy(args, cursor.SQLArguments())
 
 	rows, err := conn.Query(ctx, q, args)
@@ -204,7 +213,6 @@ INSERT INTO continual_improvement_registries (
 	organization_id,
 	reference_id,
 	description,
-	audit_id,
 	source,
 	owner_id,
 	target_date,
@@ -218,7 +226,6 @@ INSERT INTO continual_improvement_registries (
 	@organization_id,
 	@reference_id,
 	@description,
-	@audit_id,
 	@source,
 	@owner_id,
 	@target_date,
@@ -235,7 +242,6 @@ INSERT INTO continual_improvement_registries (
 		"organization_id": cir.OrganizationID,
 		"reference_id":    cir.ReferenceID,
 		"description":     cir.Description,
-		"audit_id":        cir.AuditID,
 		"source":          cir.Source,
 		"owner_id":        cir.OwnerID,
 		"target_date":     cir.TargetDate,
@@ -262,7 +268,6 @@ func (cir *ContinualImprovementRegistry) Update(
 UPDATE continual_improvement_registries SET
 	reference_id = @reference_id,
 	description = @description,
-	audit_id = @audit_id,
 	source = @source,
 	owner_id = @owner_id,
 	target_date = @target_date,
@@ -280,7 +285,6 @@ WHERE
 		"id":           cir.ID,
 		"reference_id": cir.ReferenceID,
 		"description":  cir.Description,
-		"audit_id":     cir.AuditID,
 		"source":       cir.Source,
 		"owner_id":     cir.OwnerID,
 		"target_date":  cir.TargetDate,
@@ -318,6 +322,61 @@ WHERE
 	_, err := conn.Exec(ctx, q, args)
 	if err != nil {
 		return fmt.Errorf("cannot delete continual improvement registry: %w", err)
+	}
+
+	return nil
+}
+
+func (cirs ContinualImprovementRegistries) Snapshot(ctx context.Context, conn pg.Conn, scope Scoper, organizationID, snapshotID gid.GID) error {
+	query := `
+INSERT INTO continual_improvement_registries (
+	id,
+	tenant_id,
+	snapshot_id,
+	source_id,
+	organization_id,
+	reference_id,
+	description,
+	source,
+	owner_id,
+	target_date,
+	status,
+	priority,
+	created_at,
+	updated_at
+)
+SELECT
+	generate_gid(decode_base64_unpadded(@tenant_id), @continual_improvement_registry_entity_type),
+	@tenant_id,
+	@snapshot_id,
+	r.id,
+	r.organization_id,
+	r.reference_id,
+	r.description,
+	r.source,
+	r.owner_id,
+	r.target_date,
+	r.status,
+	r.priority,
+	r.created_at,
+	r.updated_at
+FROM continual_improvement_registries r
+WHERE %s AND r.organization_id = @organization_id AND r.snapshot_id IS NULL
+	`
+
+	query = fmt.Sprintf(query, scope.SQLFragment())
+
+	args := pgx.StrictNamedArgs{
+		"tenant_id":       scope.GetTenantID(),
+		"snapshot_id":     snapshotID,
+		"organization_id": organizationID,
+		"continual_improvement_registry_entity_type": ContinualImprovementRegistryEntityType,
+	}
+	maps.Copy(args, scope.SQLArguments())
+
+	_, err := conn.Exec(ctx, query, args)
+	if err != nil {
+		return fmt.Errorf("cannot insert continual improvement registry snapshots: %w", err)
 	}
 
 	return nil