+++ pkg/coredata/entity_type_reg.go
@@ -49,4 +49,5 @@ const (
 	SnapshotEntityType
 	ContinualImprovementRegistryEntityType
 	IncidentRegistryEntityType
+	ProcessingActivityRegistryEntityType
 )