+++ pkg/coredata/snapshots_type.go
@@ -24,12 +24,14 @@ type (
 )
 
 const (
-	SnapshotsTypeRisks                   SnapshotsType = "RISKS"
-	SnapshotsTypeVendors                 SnapshotsType = "VENDORS"
-	SnapshotsTypeAssets                  SnapshotsType = "ASSETS"
-	SnapshotsTypeData                    SnapshotsType = "DATA"
-	SnapshotsTypeNonConformityRegistries SnapshotsType = "NON_CONFORMITY_REGISTRIES"
-	SnapshotsTypeComplianceRegistries    SnapshotsType = "COMPLIANCE_REGISTRIES"
+	SnapshotsTypeRisks                          SnapshotsType = "RISKS"
+	SnapshotsTypeVendors                        SnapshotsType = "VENDORS"
+	SnapshotsTypeAssets                         SnapshotsType = "ASSETS"
+	SnapshotsTypeData                           SnapshotsType = "DATA"
+	SnapshotsTypeNonConformityRegistries        SnapshotsType = "NONCONFORMITY_REGISTRIES"
+	SnapshotsTypeComplianceRegistries           SnapshotsType = "COMPLIANCE_REGISTRIES"
+	SnapshotsTypeContinualImprovementRegistries SnapshotsType = "CONTINUAL_IMPROVEMENT_REGISTRIES"
+	SnapshotsTypeProcessingActivityRegistries   SnapshotsType = "PROCESSING_ACTIVITY_REGISTRIES"
 )
 
 func (st SnapshotsType) String() string {
@@ -60,6 +62,10 @@ func (st *SnapshotsType) Scan(value any) error {
 		*st = SnapshotsTypeNonConformityRegistries
 	case SnapshotsTypeComplianceRegistries.String():
 		*st = SnapshotsTypeComplianceRegistries
+	case SnapshotsTypeContinualImprovementRegistries.String():
+		*st = SnapshotsTypeContinualImprovementRegistries
+	case SnapshotsTypeProcessingActivityRegistries.String():
+		*st = SnapshotsTypeProcessingActivityRegistries
 	default:
 		return fmt.Errorf("invalid SnapshotsType value: %q", s)
 	}